requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
appnope==0.1.4
arxiv==2.2.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.4.0
azure-common==1.1.28
azure-core==1.35.1
azure-identity==1.25.1
azure-search==1.0.0b2
azure-search-documents==11.6.0
beautifulsoup4==4.13.3
blockbuster==1.5.25
boto3==1.40.51
botocore==1.40.51
bottleneck==1.6.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
duckduckgo-search==8.1.1
ecdsa==0.19.1
exa-py==1.16.1
executing==2.2.1
fastapi==0.119.0
feedparser==6.0.12
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
google-ai-generativelanguage==0.7.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-aiplatform==1.120.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.43.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.32.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.1.0
ipykernel==7.0.0
ipython==9.6.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-rs==0.29.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-aws==0.2.35
langchain-community==0.3.31
langchain-core==0.3.79
langchain-deepseek==0.1.4
langchain-google-genai==2.1.12
langchain-google-vertexai==2.1.2
langchain-groq==0.3.8
langchain-mcp-adapters==0.1.11
langchain-openai==0.3.35
langchain-tavily==0.2.12
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-api==0.4.39
langgraph-checkpoint==2.1.2
langgraph-checkpoint-postgres==2.0.25
langgraph-cli==0.4.3
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
langsmith==0.4.34
linkup-sdk==0.6.0
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numexpr==2.14.1
numpy==1.26.4
oauthlib==3.3.1
-e file:///Users/andresb/Projects/gocareerpath
openai==2.3.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pip==25.2
platformdirs==4.5.0
pluggy==1.6.0
postgrest==2.22.0
primp==0.15.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.1
pydantic-core==2.41.3
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
realtime==2.22.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.48.0
storage3==2.22.0
strenum==0.4.15
structlog==25.4.0
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tavily-python==0.7.12
tenacity==9.1.2
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
